Mad changes in views and controllers

+++ app/controllers/likes_controller.rb
@@ -3,20 +3,14 @@ class LikesController < ApplicationController
 
   def create
     likeable = find_likeable
-    unless likeable.likes.exists?(user: current_user)
-      likeable.likes.create(user: current_user)
-      likeable.increment!(:like_count)
-    end
+    likeable.likes.create(user: current_user) unless likeable.likes.exists?(user: current_user)
     redirect_back fallback_location: root_path
   end
 
   def destroy
     likeable = find_likeable
     like = likeable.likes.find_by(user: current_user)
-    if like
-      like.destroy
-      likeable.decrement!(:like_count)
-    end
+    like.destroy if like
     redirect_back fallback_location: root_path
   end
